Do not create ProfileChange records if there is noone to notify

+++ app/services/concerns/profile_change_notifier.rb
@@ -2,13 +2,18 @@
 
 module ProfileChangeNotifier
   def prepare_profile_change(account)
+    return unless notify_profile_change_recipients(account).exists?
     ProfileChange.where(account: account).destroy_all
     @profile_change = ProfileChange.create!(account: account, avatar: account.avatar, display_name: account.display_name)
   end
 
+  def notify_profile_change_recipients(account)
+    account.followers.local
+  end
+
   def notify_profile_change
     return if @profile_change.nil?
-    @profile_change.account.followers.local.find_each do |follower|
+    notify_profile_change_recipients(@profile_change.account).find_each do |follower|
       NotifyService.new.call(follower, @profile_change)
     end
   end
